+++ web/src/pages/WishListPage/index.tsx
@@ -2,9 +2,9 @@
 import { css } from '@emotion/react';
 import { FaHeart } from 'react-icons/fa';
 import { useNavigate } from 'react-router-dom';
-import NoImage from '@/assets/images/NoImage.png';
 import emptyWishImg from '@/assets/images/empty-wish.png';
 import theme from '@/styles/theme';
+import WishListCard from './components/WishListCard';
 
 // API 명세에 맞는 타입 정의
 interface WishShelter {
@@ -39,10 +39,6 @@ const mockWishList: WishShelter[] = [
   },
 ];
 
-const handleImageError = (event: React.SyntheticEvent<HTMLImageElement>) => {
-  event.currentTarget.src = NoImage;
-};
-
 const WishListPage = () => {
   const wishList = mockWishList; // TODO: 추후 API 연결 시 변경
   const navigate = useNavigate();
@@ -62,47 +58,7 @@ const WishListPage = () => {
           </div>
           <div css={listBox}>
             {wishList.map((item) => (
-              <div
-                key={item.shelterId}
-                css={card}
-                onClick={() => handleCardClick(item.shelterId)}
-                style={{ cursor: 'pointer' }} // 클릭 가능한 UI 표시
-              >
-                <div css={cardTitleRow}>
-                  <span css={cardTitle}>{item.name}</span>
-                  <FaHeart color="red" size={30} css={cardHeart} />
-                </div>
-                <div css={cardBottomRow}>
-                  <img
-                    src={item.photoUrl && item.photoUrl.trim() !== '' ? item.photoUrl : NoImage}
-                    alt="찜 이미지"
-                    css={cardImg}
-                    onError={handleImageError}
-                  />
-                  <div css={cardInfo}>
-                    <div css={cardRating}>
-                      별점: <span css={ratingNumber}>{item.averageRating}</span>
-                      <span css={starsWrapper}>
-                        {Array.from({ length: 5 }, (_, i) => (
-                          <span
-                            key={i}
-                            css={i < Math.round(item.averageRating) ? filledStar : emptyStar}
-                          >
-                            ★
-                          </span>
-                        ))}
-                      </span>
-                    </div>
-                    <div css={cardinfostyle}>
-                      거리: {item.distance}
-                      <br />
-                      운영시간: {item.operatingHours}
-                      <br />
-                      주소: {item.address}
-                    </div>
-                  </div>
-                </div>
-              </div>
+              <WishListCard key={item.shelterId} item={item} onClick={handleCardClick} />
             ))}
           </div>
         </div>
@@ -161,96 +117,6 @@ const listBox = css`
   flex-direction: column;
   gap: 12px;
   width: 95%;
-`;
-
-const card = css`
-  background: #e0e0e0;
-  border-radius: 12px;
-  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.15);
-  overflow: hidden;
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  padding: 12px 0;
-`;
-
-const cardTitleRow = css`
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  gap: 8px;
-  padding: 0 16px 8px 16px;
-`;
-
-const cardTitle = css`
-  ${theme.typography.wish2};
-  padding-bottom: 4px;
-`;
-
-const cardHeart = css`
-  font-size: ${theme.typography.wish1};
-  padding-bottom: 4px;
-`;
-
-const cardBottomRow = css`
-  display: flex;
-  flex-direction: row;
-  align-items: flex-start;
-  gap: 12px;
-  padding: 0 16px;
-`;
-
-const cardImg = css`
-  width: 30%;
-  height: 30%;
-  object-fit: cover;
-  border-radius: 8px;
-  background: #fafafa;
-`;
-
-const cardInfo = css`
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  justify-content: flex-start;
-`;
-
-const cardRating = css`
-  display: flex;
-  align-items: center;
-  gap: 3px;
-  ${theme.typography.wish3};
-  color: ${theme.colors.text.gray500};
-`;
-
-const ratingNumber = css`
-  ${theme.typography.wish3};
-  color: ${theme.colors.text.red};
-`;
-
-const starsWrapper = css`
-  display: inline-flex;
-  align-items: center;
-  gap: 2px;
-  ${theme.typography.wish3};
-  color: ${theme.colors.text.gray500};
-`;
-
-const filledStar = css`
-  color: #ffd700;
-  ${theme.typography.wish3};
-`;
-
-const emptyStar = css`
-  color: #bbb;
-  ${theme.typography.wish3};
-`;
-
-const cardinfostyle = css`
-  text-align: left;
-  ${theme.typography.wish3};
-  color: ${theme.colors.text.gray500};
 `;
 
 const emptyStateStyle = css`
